+++ include/ensmallen_bits/ens_version.hpp
@@ -15,17 +15,17 @@
 #define ENS_VERSION_MAJOR 2
 // The minor version is two digits so regular numerical comparisons of versions
 // work right.  The first minor version of a release is always 10.
-#define ENS_VERSION_MINOR 18
-#define ENS_VERSION_PATCH 2
+#define ENS_VERSION_MINOR 19
+#define ENS_VERSION_PATCH 0
 // If this is a release candidate, it will be reflected in the version name
 // (i.e. the version name will be "RC1", "RC2", etc.).  Otherwise the version
 // name will typically be a seemingly arbitrary set of words that does not
 // contain the capitalized string "RC".
-#define ENS_VERSION_NAME "Fairmount Bagel"
+#define ENS_VERSION_NAME "Eight Ball Deluxe"
 // Incorporate the date the version was released.
 #define ENS_VERSION_YEAR "2022"
-#define ENS_VERSION_MONTH "02"
-#define ENS_VERSION_DAY "13"
+#define ENS_VERSION_MONTH "04"
+#define ENS_VERSION_DAY "06"
 
 namespace ens {